\name{test88-package}
\alias{test88-package}
\alias{test88}
\docType{package}
\title{
\packageTitle{test88}
}
\description{
\packageDescription{test88}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{test88}
\packageIndices{test88}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{test88}

Maintainer: \packageMaintainer{test88}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
